introduce RoundBased UseCase
This takes care of setup stuff (database, etc.). Implementing
subclasses can just focus upon their "business logic" by using the
callbacks: `setup`, `perform_round` and `teardown`.
